+++ old/lab-4/src/index.js
@@ -1,0 +1,242 @@
+import React, { useState } from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
+import App from "./App.js";
+
+function Square(props) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
+
+class Board extends React.Component {
+  renderSquare(i) {
+    return (
+      <Square
+        value={this.props.squares[i]}
+        onClick={() => {
+          this.props.onClick(i);
+        }}
+      />
+    );
+  }
+
+  render() {
+    return (
+      <div>
+        <div className="board-row">
+          {this.renderSquare(0)}
+          {this.renderSquare(1)}
+          {this.renderSquare(2)}
+        </div>
+        <div className="board-row">
+          {this.renderSquare(3)}
+          {this.renderSquare(4)}
+          {this.renderSquare(5)}
+        </div>
+        <div className="board-row">
+          {this.renderSquare(6)}
+          {this.renderSquare(7)}
+          {this.renderSquare(8)}
+        </div>
+      </div>
+    );
+  }
+}
+
+class Game extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      history: [
+        {
+          squares: Array(9).fill(null),
+        },
+      ],
+      stepNumber: 0,
+      xIsNext: true,
+    };
+  }
+
+  handleClick(i) {
+    const history = this.state.history.slice(0, this.state.stepNumber + 1);
+    const current = history[history.length - 1];
+    const squares = current.squares.slice();
+    if (calculateWinner(squares) || squares[i]) {
+      return;
+    }
+    squares[i] = this.state.xIsNext ? "X" : "O";
+    this.setState({
+      history: history.concat([
+        {
+          squares: squares,
+        },
+      ]),
+      stepNumber: history.length,
+      xIsNext: !this.state.xIsNext,
+    });
+  }
+
+  jumpTo(step) {
+    this.setState({
+      stepNumber: step,
+      xIsNext: step % 2 === 0,
+    });
+  }
+
+  render() {
+    const history = this.state.history;
+    const current = history[this.state.stepNumber];
+    const winner = calculateWinner(current.squares);
+
+    const moves = history.map((step, move) => {
+      const desc = move ? "Go to move # " + move : "Go to game start";
+
+      return (
+        <li key={move}>
+          <button onClick={() => this.jumpTo(move)}>{desc}</button>
+        </li>
+      );
+    });
+
+    let status;
+    if (winner) {
+      status = "Winner: " + winner;
+    } else {
+      status = "Next player: " + (this.state.xIsNext ? "X" : "O");
+    }
+
+    return (
+      <div style={{margin: "25px"}}>
+        <h4>
+          Tic-Tac-Toe game from{" "}
+          <a href="https://reactjs.org/tutorial/tutorial.html">
+            React Intro Tutorial
+          </a>
+        </h4>
+        <div className="game">
+          <div className="game-board">
+            <Board
+              squares={current.squares}
+              onClick={(i) => this.handleClick(i)}
+            />
+          </div>
+          <div className="game-info">
+            <div>{status}</div>
+            <ol>{moves}</ol>
+          </div>
+        </div>
+      </div>
+    );
+  }
+}
+
+function calculateWinner(squares) {
+  const lines = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [2, 4, 6],
+  ];
+  for (let i = 0; i < lines.length; i++) {
+    const [a, b, c] = lines[i];
+    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+      return squares[a];
+    }
+  }
+  return null;
+}
+
+// ========================================
+
+function Messenger(props) {
+  const [message, setMessage] = useState("");
+  const [messages, setMessages] = useState([]);
+
+  const handleSubmit = (event) => {
+    event.preventDefault();
+
+    let date = new Date(); // date will be determined server-side in the future to prevent forgery
+    date = date.toLocaleTimeString();
+    setMessages([...messages, { date, message }]);
+
+    setMessage(""); // reset message to clear the input field
+  };
+
+  const listMessages = messages.map((msg) => (
+    <div>
+      <span
+        style={{
+          fontSize: "75%",
+          marginLeft: "5px",
+        }}
+      >
+        {msg.date} ={"> "}
+      </span>
+      <span>{msg.message}</span>
+    </div>
+  ));
+
+  return (
+    <div style={{margin: "20px"}}>
+      <br></br>
+      <hr></hr>
+      <h4>Simple Messenger using functionnal components and hooks</h4>
+
+      <div
+        style={{
+          border: "solid 1px black",
+          marginLeft: "5px",
+          marginRight: "50%",
+          height: "200px",
+        }}
+      >
+        {listMessages}
+      </div>
+
+      <form onSubmit={handleSubmit} style={{ margin: "5px" }}>
+        <label>
+          <span style={{ margin: "5px" }}>Speak:</span>
+          <input
+            type="text"
+            id="field"
+            value={message}
+            onChange={(e) => setMessage(e.target.value)}
+          />
+        </label>
+        <input type="submit" value="Submit" style={{ margin: "5px" }} />
+      </form>
+      <br></br>
+      <Link to="/App">
+        Another example using DOM modification instead of state modification for
+        input field reset + CSS-in-JS w/ Emotion
+      </Link>
+    </div>
+  );
+}
+
+// ========================================
+
+ReactDOM.render(
+  <Router>
+    <div>
+      <Switch>
+        <Route exact path="/">
+          <Game />
+          <Messenger />
+        </Route>
+        <Route exact path="/App">
+          <App />
+        </Route>
+      </Switch>
+    </div>
+  </Router>,
+  document.getElementById("root")
+);