+++ CHANGELOG.md
@@ -6,8 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased](https://github.com/benzinho75/node-messenger/compare/v0.0.3...HEAD)
+## [Unreleased](https://github.com/benzinho75/node-messenger/compare/v0.0.4...HEAD)
 
+
+## [0.0.4](https://github.com/benzinho75/node-messenger/compare/v0.0.3...v0.0.4) - ?
+
+### Added
+- Lab 4: other messenger example using DOM modification instead of state modification for input field reset + CSS-in-JS w/ Emotion
+- Lab 4: declarative routing with React Router
+- Refactored project structure
+- Refactored front-end in smaller components
 
 ## [0.0.3](https://github.com/benzinho75/node-messenger/compare/v0.0.2...v0.0.3) - 2020-10-14
 
@@ -25,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CHANGELOG.md links
 - Lab 3 (WIP: 7 out of 9 tests passed)
-
 
 ### Fixed
 - README.md